pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tdrcreator"
version = "0.1.0"
description = "Local-only Transfer Documentation Report generator with private RAG"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "typer[all]>=0.12",
    "pyyaml>=6.0",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    "numpy>=1.26",
    "requests>=2.31",
    "rich>=13.0",
    "jinja2>=3.1",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3", "Pillow>=10.0"]
pdf-export = ["reportlab>=4.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "responses>=0.25",
]

[project.scripts]
tdrcreator = "tdrcreator.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["tdrcreator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
